05/src/pyproject.toml
[project]
name = "yaml-json-converter"
version = "0.1.0"
description = "YAML to JSON Converter"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask[async]>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "argon2-cffi>=23.1.0",
    "pyyaml>=6.0.0",
    "gunicorn>=23.0.0",
    "gevent>=25.9.1",
]
